requirements.txt
streamlit
firebase-admin
opencv-python-headless
requests
numpy
pillow